roompage以下のtitle一時設定とh1の変更。ステ編集ページの追加
