+++ the_project/grouppage/templates/grouppage/roompage.html
@@ -7,7 +7,7 @@
         <meta http-equiv="refresh" content="60; URL=">
     </head>
     <body>
-    <h1>{{groupname}}です</h1>
+    <h1>{{projectname}}/{{groupname}}</h1>
 
     ディスプレイネーム:{{displayname_role.displayname}}<br>
     役職:{{displayname_role.role}}<br><br>
@@ -48,7 +48,7 @@
 
 
 
-    <a href="">グループ設定</a>
+    <a href="/mainpage/project_{{projectid}}/grouppage_{{group.uuid}}/group_settings">グループ設定</a>
     <a href="/mainpage/project_{{projectid}}">プロジェクトページに戻る</a>
 
     <br><br><br>
