+++ the_project/grouppage/views.py
@@ -1,5 +1,6 @@
 from .view.roompage import *
-from .view.status_new import *
+from .view.status_edit import *
+from .view.group_settings import *
 
 #roompage→部屋ページ
 #status_new→新しいステータス作るとき
