+++ the_project/mainpage/templates/mainpage/grouppage.html
@@ -46,8 +46,4 @@
     {% endfor %}
 </table>
 -->
-
-
-
-
 {% endblock %}
